\name{cp.types}
\alias{cp.types}
\docType{data}
\title{
Probe types
}

\description{
Vector of types of control probes to be used in examples for funtoonorm function
or if cp.type variable is not provided by user.
}
\usage{cp.types}

\format{
  Vector of length 835 (one for each control probe), each element is a
  character string describing the type of control probe.
}
