% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\name{fromGenStudFiles}
\alias{fromGenStudFiles}
\title{create a sample set from GenomeStudio files}
\usage{
fromGenStudFiles(controlProbeFile, signalFile, cell_type)
}
\arguments{
\item{controlProbeFile}{file exported from GenomeStudio}

\item{signalFile}{file exported from GenomeStudio with the exact same samples}

\item{cell_type}{vector should have names matching the 2 files from genome
studios, and at least 2 cell types.}
}
\value{
a SampleSet object
}
\description{
create a sample set from GenomeStudio files
}

