\name{funtooNorm}
\alias{funtooNorm}
\docType{package}
\title{
NewNorm
}
\description{Package provides a function to normalize Illumina Infinium Human Methylation 450 BeadChip (Illumina 450K), correcting for  tissue and/or cell type.
}
\details{
\tabular{ll}{
Package: \tab funtooNorm\cr
Type: \tab Package\cr
Version: \tab 0.99.1\cr
Date: \tab 2015-04-29\cr
License: \tab GPL-3\cr
}
}
%%\author{Celia Greenwood, Stepan Grinek}

%\references{
%Literature or other references for background information
%}

%%\keyword{Methylation, Preprocessing, PLS}

%%\examples{
%%    data("data")
%%    ncmp <- 4
%%   funtoonormout <- funtoonorm(sigA=sigAsample, sigB=sigBsample, Annot=Annotsample, 
%%                      controlred=matred, controlgrn=matgrn, 
%%                      cp.types=cp.types, cell_type = cell_type,
%%                       ncmp=ncmp, save.quant=TRUE, save.loess=TRUE, apply.loess=TRUE, logit.quant=TRUE, validate=FALSE)
    
%%}
