% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\docType{methods}
\name{funtooNorm}
\alias{funtooNorm}
<<<<<<< HEAD
\alias{funtooNorm,SampleSet-method}
\title{The funtooNorm normalization function}
=======
\title{This is the main normalization function which applies to autosomes and the X chromosome.  
Chromosome Y requires separate analysis as there are few probes on Y.  
We use a straightforward quantile normalization applied to males only.}

>>>>>>> 9e15e6d42974a0953ef22dc78fb91d6d4a9f81ef
\usage{
funtooNorm(object, type.fits = "PCR", ncmp = 4, force = FALSE,
  sex = NULL)

\S4method{funtooNorm}{SampleSet}(object, type.fits = "PCR", ncmp = 4,
  force = FALSE, sex = NULL)
}
\arguments{
\item{object}{Object of class SampleSet}

\item{type.fits}{Choice between "PCR" or "PLS" (default="PCR")}

\item{ncmp}{Number of components included in the analysis (default=4)}

\item{force}{When set to TRUE forces the normalization procedure to re-compute}

\item{sex}{Boolean vector if male. if NULL Beta values from ChrY are used for
classification.}
}
\value{
a S4 object of class SampleSet containing the normalized signal
}
\description{
\code{funtooNorm}
 
This is a generic function which applies to autosomes and the X chromosome.  
Chromosome Y requires separate analysis as there are few probes on Y.  
We use a straightforward quantile normalization applied to males only.
}
\section{Methods (by class)}{
\itemize{
\item \code{SampleSet}: The funtooNorm normalization function
}}
\examples{
require(minfiData)
pData(RGsetEx)$cell_type <- rep(c("type1","type2"),3)
mySampleSet=fromRGChannelSet(RGsetEx)
mySampleSet=funtooNorm(mySampleSet)

}

