% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\name{getAnnotation}
\alias{getAnnotation}
\title{Get the SAmpleSet Annotation in the same order as the beta matrix}
\usage{
getAnnotation(object)
}
\arguments{
\item{object}{of type SampleSet}
}
\value{
annotation
}
\description{
Get the SAmpleSet Annotation in the same order as the beta matrix
}

