% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\name{getLogSigA}
\alias{getLogSigA}
\title{internal function to get the signal A}
\usage{
getLogSigA(signal)
}
\arguments{
\item{signal}{}
}
\value{
all signal A in the good order
}
\description{
internal function to get the signal A
}

