% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\name{getLogSigB}
\alias{getLogSigB}
\title{internal function to get the signal B}
\usage{
getLogSigB(signal)
}
\arguments{
\item{signal}{}
}
\value{
all signal B in the good order
}
\description{
internal function to get the signal B
}

