% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\name{getNormBeta}
\alias{getNormBeta}
\title{compute the beta value after normalization for each position and each sample}
\usage{
getNormBeta(object, offset = 100)
}
\arguments{
\item{object}{of type SampleSet}

\item{offset}{default is 100 as Illumina standard}
}
\value{
a matrix containing beta after normalization value for each CpG 
position and each samples
}
\description{
compute the beta value after normalization for each position and each sample
}
\examples{
require(minfiData)
mySampleSet=fromRGChannelSet(RGsetEx)
getNormBeta(mySampleSet)
}

