% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\name{getRawBeta}
\alias{getRawBeta}
\title{compute the beta value of the raw signal for each position and each sample}
\usage{
getRawBeta(object, offset = 100)
}
\arguments{
\item{object}{of type SampleSet}

\item{offset}{default is 100 as Illumina standard}
}
\value{
a matrix containing the raw beta value for each position and each
samples
}
\description{
compute the beta value of the raw signal for each position and each sample
}
\examples{
require(minfiData)
mySampleSet=fromRGChannelSet(RGsetEx)
getRawBeta(mySampleSet)

}

