% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\name{getSampleSetAnnotation}
\alias{getSampleSetAnnotation}
\title{Get the SampleSet Annotation in the same order as the beta matrix}
\usage{
getSampleSetAnnotation()
}
\arguments{
\item{object}{of type SampleSet}
}
\value{
annotation
}
\description{
Get the SampleSet Annotation in the same order as the beta matrix
}
\examples{
getSampleSetAnnotation(mySampleSet)
}

