% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\name{getSnpM}
\alias{getSnpM}
\title{Computes the M value after normalization for each SNP.}
\usage{
getSnpM(object)
}
\arguments{
\item{object}{of class SampleSet}
}
\value{
a matrix containing M values, log2(Meth/Unmeth), after normalization for each SNP
}
\description{
Computes the M value after normalization for each SNP.
}
\examples{
require(minfiData)
pData(RGsetEx)$cell_type <- rep(c("type1","type2"),3)
mySampleSet=fromRGChannelSet(RGsetEx)
snp=getSnpM(funtooNorm(mySampleSet))

}

