\name{individualID}
\alias{individualID}
\docType{data}
\title{
Individual ID's
}
\description{
Vector of individual IDs, for evaluation of the performance of funtoonorm when using agreement function.
}
\usage{individualID}

\format{
  A vector of length 93, containing the individual IDs. The sample names are
  used to name the vector elements.
}
