\name{matgrn}
\alias{matgrn}
\docType{data}
\title{
Green control signals
}
\description{
Matrix containing the control probe signals from the green channel.  The column names should be the same as in the signal A and signal B matrices.}
\usage{matgrn}

\format{
  A matrix with 835 rows (one for each control probe) and 93 columns
  (one per sample). Each element is the control probe signal for the
  given probe and sample.
}
