\name{matred}
\alias{matred}
\docType{data}
\title{
Red control signals
}
\description{
Matrix containing the control probe signals from the red channel.  The column names should be the same as in the signal A and signal B matrices.
}
\usage{matred}

\format{
  A matrix with 835 rows (one for each control probe) and 93 columns
  (one per sample). Each element is the control probe signal for the
  given probe and sample.
}
