% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\name{plotValidationGraph}
\alias{plotValidationGraph}
\title{Plot a series of graphs for each signal type, illustrating the effects
 of the number of components included}
\usage{
plotValidationGraph(object, type.fits = "PCR", file = NULL)
}
\arguments{
\item{object}{of class SampleSet}

\item{type.fits}{can be "PCR" or "PLS" (default "PCR")}

\item{file}{if not NULL will write a pdf using this name, path can be
included}
}
\description{
Plot a series of graphs for each signal type, illustrating the effects
 of the number of components included
}
\examples{
require(minfiData)
pData(RGsetEx)$cell_type <- rep(c("type1","type2"),3)
mySampleSet=fromRGChannelSet(RGsetEx)
plotValidationGraph(mySampleSet)

}

