% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\docType{class}
\name{SampleSet-class}
\alias{SampleSet-class}
\title{SampleSet are S3 objects define for the purpose of runing funtooNorm
algorithm, they are list containing signal data and different variables
usefull for funtooNorm. The data is separated into the 3 probes type, having
2 channels (methylated and unmethylated ie : A and B.
We then define then the 6 (2*3) labels: AIGrn BIGrn AIRed BIRed AII BII}
\value{
a SampleSet object
}
\description{
SampleSet are S3 objects define for the purpose of runing funtooNorm
algorithm, they are list containing signal data and different variables
usefull for funtooNorm. The data is separated into the 3 probes type, having
2 channels (methylated and unmethylated ie : A and B.
We then define then the 6 (2*3) labels: AIGrn BIGrn AIRed BIRed AII BII
}
\section{Slots}{

\describe{
\item{\code{type}}{character: is 'minfi' or 'GenomeStudio'}

\item{\code{sampleNames}}{character vector:
contain the list of sample names in order used}

\item{\code{sampleSize}}{numeric: the number of samples}

\item{\code{nPos}}{numeric: the number of positions in the ILLUMINA chip}

\item{\code{cell_type}}{list: list matching each sample to define the categories}

\item{\code{qntllist}}{numeric: vector of ordered quantiles}

\item{\code{quantiles}}{numeric: list of  6 auqntiles tables for 6 type of signals}

\item{\code{ctl.covmat}}{numeric: covariance matrix for the model fit}

\item{\code{signal}}{numeric: list of 6 signal tables the 6 type of signals}
}}
\examples{
showClass("SampleSet")

}

