% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\docType{methods}
\name{show,SampleSet-method}
\alias{show,SampleSet-method}
\title{Show Object SampleSet}
\usage{
\S4method{show}{SampleSet}(object)
}
\arguments{
\item{object}{an object of class SampleSet}

\item{...}{optional arguments passed to or from other methods.}
}
\value{
No value is returned. The function prints the summary of object of 
class SampleSet to screen
}
\description{
Display informations about the SampleSet object
}
\examples{
require(minfiData)
pData(RGsetEx)$cell_type <- rep(c("type1","type2"),3)
mySampleSet=fromRGChannelSet(RGsetEx)
mySampleSet

}

