\name{sigAsample}
\alias{sigAsample}
\docType{data}
\title{
Signal A
}
\description{
A matrix containing the signal A results from the Illumina Infinium HumanMethylation450 BeadChip, with columns corresponding to the samples and rows to the probes.  The sample data provided contains a selection of only 20,0000 probes for demonstrating the funtooNorm function. }
\usage{sigAsample}
