\name{sigBsample}
\alias{sigBsample}
\docType{data}
\title{
Signal B
}
\description{
A matrix containing the signal B results from the Illumina Infinium HumanMethylation450 BeadChip, with columns corresponding to the samples and rows to the probes.  The sample data provided contains a selection of only 20,0000 probes for demonstrating the funtooNorm function.
}
\usage{sigBsample}


