% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funtoonorm.R
\name{agreement}
\alias{agreement}
\title{Function to measure intra-replicate agreement for methylation data.}
\usage{
agreement(Beta, individualID)
}
\arguments{
\item{Beta}{: Matrix with beta-values, rows corresponding to probes, columns
corresponding to samples.}

\item{individualID}{: a vector where 2 replicates have the exact same value
for two technical replicates. Order of samples should nmatch the samples
(columns) in Beta}
}
\value{
The average value of the square distance between replicates: 
a measure of agreement between replicates in methylation data.
}
\description{
Function to measure intra-replicate agreement for methylation data.
}
\details{
We expect that the values returned by the agreement function after
 normalization by funtooNorm to be smaller than before.
}
\examples{
agreement(cbind(rnorm(n = 10),rnorm(n = 10),rnorm(n = 10)),c(1,1,1))

}

