% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\name{fromGenStudFiles}
\alias{fromGenStudFiles}
\title{Creates a S4 object of class 'SampleSet' from GenomeStudio files}
\usage{
fromGenStudFiles(controlProbeFile, signalFile, cell_type)
}
\arguments{
\item{controlProbeFile}{file of control probe data exported from GenomeStudio}

\item{signalFile}{file exported from GenomeStudio with the exact same samples
as control probe File}

\item{cell_type}{this vector should have names matching all the samples in
the files from genome studios, and at least 2 different cell types.}
}
\value{
a SampleSet object
}
\description{
Creates a S4 object of class 'SampleSet' from GenomeStudio files
}

