% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\name{fromGenStudFiles}
\alias{fromGenStudFiles}
\title{Creates a S4 object of class 'SampleSet' from GenomeStudio files}
\usage{
fromGenStudFiles(controlProbeFile, signalFile, cell_type)
}
\arguments{
\item{controlProbeFile}{The control probe file exported from GenomeStudio}

\item{signalFile}{The signals exported from GenomeStudio samples must be in
same order as the control probe File}

\item{cell_type}{A vector of cell types, names must match control probes and
signal files.}
}
\value{
An object of class 'SampleSet'.
}
\description{
Creates a S4 object of class 'SampleSet' from GenomeStudio files
}

