\name{funtooNorm-package}
\alias{funtooNorm-package}
\docType{package}
\title{
funtooNorm
}
\description{The funtooNorm Package provides a normalization method for data
arising from the Illumina Infinium Human Methylation 450 BeadChip 
(Illumina 450K), including explicit considerations of differences between
tissues or cell types.  This method should only be used when the data set
contains samples fom multiple different tissues or cell types.
}
\details{
\tabular{ll}{
Package: \tab funtooNorm\cr
Type: \tab Package\cr
Version: \tab 0.99.3\cr
Date: \tab 2016-03-28\cr
License: \tab GPL-3\cr
}
}
\author{Celia Greenwood, Raphael Poujol, Stepan Grinek, Kathleen Oros Klein}

%\references{
%Literature or other references for background information
%}

\keyword{Methylation, Preprocessing, PLS}

