\name{funtooNorm-package}
\alias{funtooNorm-package}
\docType{package}
\title{
funtooNorm
}
\description{The funtooNorm Package provides a normalization method for data
arising from the Illumina Infinium Human Methylation 450 BeadChip 
(Illumina 450K), including explicit considerations of differences between
tissues or cell types.  This method should only be used when the data set
contains samples fom multiple different tissues or cell types.
}
\details{
\tabular{ll}{
Package: \tab funtooNorm\cr
Type: \tab Package\cr
License: \tab GPL-3\cr
}
}
\author{Celia Greenwood, Stepan Grinek, Raphael Poujol, Maxime Turgeon, 
Kathleen Oros Klein}

%\references{
%Literature or other references for background information
%}

\keyword{Methylation, Preprocessing, PLS}

