% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\name{funtooNorm}
\alias{funtooNorm}
\title{This is the main normalization function which applies to autosomes and the X chromosome.  
Chromosome Y requires separate analysis, since there are few probes on Y.  
Therefore we use a straightforward quantile normalization applied to males only.}
\usage{
funtooNorm(object, type.fits = "PCR", ncmp = 4, force = FALSE,
  sex = NULL)
}
\arguments{
\item{object}{of type SampleSet}

\item{type.fits}{can be "PCR" or "PLS" (default "PCR")}

\item{ncmp}{number of components used in the analysis (default 4)}

\item{force}{set it to TRUE in order to re-compute the normalisation whent 
it is already done}

\item{sex}{boolean vector: when not null force the chrY normalization to use 
treat the TRUE values as mens}
}
\value{
a SampleSet containing the normalised signal
}
\description{
This is the main normalization function which applies to autosomes and the X chromosome.  
Chromosome Y requires separate analysis, since there are few probes on Y.  
Therefore we use a straightforward quantile normalization applied to males only.
}
\examples{
require(minfiData)
pData(RGsetEx)$cell_type <- rep(c("type1","type2"),3)
mySampleSet=fromRGChannelSet(RGsetEx)
mySampleSet=funtooNorm(mySampleSet)

}

