% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\name{funtooNorm}
\alias{funtooNorm}
\title{This is the main normalization function which applies to autosomes and the X chromosome.  
Chromosome Y requires separate analysis as there are few probes on Y.  
We use a straightforward quantile normalization applied to males only.}

\usage{
funtooNorm(object, type.fits = "PCR", ncmp = 4, force = FALSE,
  sex = NULL)
}
\arguments{
\item{object}{of class SampleSet}

\item{type.fits}{can be "PCR" or "PLS" (default "PCR")}

\item{ncmp}{number of components used in the analysis (default 4)}

\item{force}{when set to TRUE forces the normalization procedure to re-compute}

\item{sex}{boolean vector: if NULL Beta values from ChrY are used for classification.}
}
\value{
a S4 object of class SampleSet containing the normalized signal
}
\description{
This is the main normalization function which applies to autosomes and the X chromosome.  
Chromosome Y requires separate analysis as there are few probes on Y.  
We use a straightforward quantile normalization applied to males only.
}
\examples{
require(minfiData)
pData(RGsetEx)$cell_type <- rep(c("type1","type2"),3)
mySampleSet=fromRGChannelSet(RGsetEx)
mySampleSet=funtooNorm(mySampleSet)

}

