% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\name{funtooNorm}
\alias{funtooNorm}
\title{This function apply the normalization method central to the package on each
signal}
\usage{
funtooNorm(object, type.fits = "PCR", ncmp = 4, force = FALSE)
}
\arguments{
\item{object}{of type SampleSet}

\item{type.fits}{can be "PCR" or "PLS" (default "PCR")}

\item{ncmp}{number of component use for anaysis (default 4)}

\item{force}{tu re-compute the normalisation whent it is already done}
}
\value{
a SampleSet containing the normalised signal
}
\description{
This function apply the normalization method central to the package on each
signal
}

