% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\docType{methods}
\name{getGRanges}
\alias{getGRanges}
\alias{getGRanges,SampleSet-method}
\title{Build GRange object of methylation probes}
\usage{
getGRanges(object)

\S4method{getGRanges}{SampleSet}(object)
}
\arguments{
\item{object}{Object of class SampleSet.}
}
\value{
A GRange object of the positions of each cpg.
}
\description{
Build GRange object of methylation probes
}
\section{Methods (by class)}{
\itemize{
\item \code{SampleSet}: Build GRange object of methylation probes
}}
\examples{
require(minfiData)
pData(RGsetEx)$cell_type <- rep(c("type1","type2"),3)
mySampleSet=fromRGChannelSet(RGsetEx)
gr=getGRanges(mySampleSet)

}

