% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\name{getNormM}
\alias{getNormM}
\title{compute the M value after normalization for each position and each sample}
\usage{
getNormM(object, offset = 100)

getNormM(object, offset = 100)
}
\arguments{
\item{object}{of type SampleSet}

\item{offset}{default is 100 as Illumina standard}

\item{object}{of type SampleSet}

\item{offset}{default is 100 as Illumina standard}
}
\value{
a matrix containing M after normalization value for each position
and each samples log(Meth/Unmeth)

a matrix containing M after normalization value for each position
and each samples
}
\description{
compute the M value after normalization for each position and each sample

compute the beta value after normalization for each position and each sample
}

