% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\name{getNormM}
\alias{getNormM}
\title{Computes the M value after normalization at each position for each
sample}
\usage{
getNormM(object)
}
\arguments{
\item{object}{of type SampleSet}

\item{offset}{default is 100 as Illumina standard}
}
\value{
a matrix containing M values, log2(Meth/Unmeth), after normalization at each position
for each samples.
}
\description{
Computes the M value after normalization at each position for each
sample
}
\examples{
require(minfiData)
pData(RGsetEx)$cell_type <- rep(c("type1","type2"),3)
mySampleSet=fromRGChannelSet(RGsetEx)
m=getNormM(funtooNorm(mySampleSet))

}

