% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\docType{methods}
\name{getNormM}
\alias{getNormM}
\alias{getNormM,SampleSet-method}
\title{Computes M values,log2(Meth/Unmeth), from normalized signals}
\usage{
getNormM(object)

\S4method{getNormM}{SampleSet}(object)
}
\arguments{
\item{object}{An object of class SampleSet}
}
\value{
a matrix containing M values, log2(Meth/Unmeth), after normalization
}
\description{
Computes M values,log2(Meth/Unmeth), from normalized signals
}
\section{Methods (by class)}{
\itemize{
\item \code{SampleSet}: Computes M values, log2(Meth/Unmeth),
from normalized signals
}}
\examples{
require(minfiData)
pData(RGsetEx)$cell_type <- rep(c("type1","type2"),3)
mySampleSet=fromRGChannelSet(RGsetEx)
m=getNormM(funtooNorm(mySampleSet))

}

