% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\name{getSnpM}
\alias{getSnpM}
\title{compute the M value after normalization for each SNP position and each sample}
\usage{
getSnpM(object)
}
\arguments{
\item{object}{of type SampleSet}

\item{offset}{default is 100 as Illumina standard}
}
\value{
a matrix containing M after normalization value for each SNP of the 
chip and each sample log2(Meth/Unmeth)
}
\description{
compute the M value after normalization for each SNP position and each sample
}
\examples{
require(minfiData)
mySampleSet=fromRGChannelSet(RGsetEx)
getSnpM(mySampleSet)

}

