% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\name{plotValidationGraph}
\alias{plotValidationGraph}
\title{Plot a series of graphs with different numbers of components for
each signal}
\usage{
plotValidationGraph(object, type.fits = "PCR", file = "")
}
\arguments{
\item{object}{of type SampleSet}

\item{type.fits}{can be "PCR" or "PLS" (default "PCR")}

\item{file}{if not empty will write a pdf using this name, path can be
included}
}
\description{
Plot a series of graphs with different numbers of components for
each signal
}
\examples{
require(minfiData)
pData(RGsetEx)$cell_type <- rep(c("type1","type2"),3)
mySampleSet=fromRGChannelSet(RGsetEx)
plotValidationGraph(mySampleSet)

}

