% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\name{print.SampleSet}
\alias{print.SampleSet}
\title{Print information about the SampleSet}
\usage{
\method{print}{SampleSet}(x, ...)
}
\arguments{
\item{x}{an object of class SampleSet}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print information about the SampleSet
}
\examples{
require(minfiData)
pData(RGsetEx)$cell_type <- rep(c("type1","type2"),3)
mySampleSet=fromRGChannelSet(RGsetEx)
mySampleSet

}

