% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSet.R
\docType{class}
\name{SampleSet-class}
\alias{SampleSet-class}
\title{SampleSet are S3 objects define for the purpose of runing funtooNorm
algorithm}
\description{
SampleSet are S3 objects define for the purpose of runing funtooNorm
algorithm
}
\section{Slots}{

\describe{
\item{\code{type}}{character: is 'minfi' or 'GenomeStudio'}

\item{\code{sampleNames}}{character vector:
contain the list of sample names in the order used}

\item{\code{sampleSize}}{numeric: the number of samples}

\item{\code{nPos}}{numeric: the number of positions in the ILLUMINA chip}

\item{\code{cell_type}}{list: list matching each sample to define the categories}

\item{\code{qntllist}}{numeric: vector of ordered quantiles}

\item{\code{quantiles}}{numeric: list of  6 auqntiles tables for 6 type of signals}

\item{\code{ctl.covmat}}{numeric: covariance matrix for the model fit}

\item{\code{signal}}{numeric: list of 6 signal tables the 6 type of signals}
}}

